requirements.txt
datapackage-pipelines
datapackage-pipelines-aws
tableschema-elasticsearch
elasticsearch
